requirements.txt
flask>=2.2
requests>=2.28
python-dotenv>=0.21
numpy>=1.24
pandas>=1.5
matplotlib>=3.6
plotly>=5.10
scikit-learn>=1.1
prophet>=1.1
pystan>=2.19
cmdstanpy>=1.0